Fix depth preview float arithmetic precision
Cast clipped depth to float32 and use float literals in division to ensure
proper floating-point math when normalizing depth values to 0-255 range.

+++ capture.py
@@ -176,7 +176,7 @@ def depth_to_preview(depth_mm: np.ndarray) -> Image.Image:
     far = int(np.percentile(valid, 95))
     if far <= near:
         far = near + 1
-    clipped = np.clip(depth_mm, near, far)
-    preview = ((clipped - near) * 255 / (far - near)).astype(np.uint8)
+    clipped = np.clip(depth_mm, near, far).astype(np.float32)
+    preview = ((clipped - near) * (255.0 / (far - near))).astype(np.uint8)
     preview[depth_mm == 0] = 0
     return Image.fromarray(preview, mode="L")
